requirements.txt
fastapi
uvicorn
ffmpeg-python
git+https://github.com/openai/whisper.git